server/pyproject.toml
[tool.poetry]
name = "movie_reviews_backend"
version = "0.1.0"
description = "server for the movie reviews site"
authors = ["Anton Kueltz <kueltz.anton@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.2"
pymongo = "^3.11.1"
python-multipart = "^0.0.5"
argon2-cffi = "^20.1.0"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]
uvicorn = "^0.12.3"
gunicorn = "^20.0.4"
uvloop = "^0.14.0"
httptools = "^0.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
